fix: _zero placeholder in mod tuple now uses compute dtype
Latent trap flagged by Task 5 review: mx.zeros(...) defaulted to float32
regardless of the model's dtype. Currently unread (only the reference's
unported fused-kernel branch consumes it), but would silently hand a
wrong-dtype tensor to whoever wires that branch up later.

+++ apple_silicon_nodes/native/qwen_image21/model.py
@@ -319,7 +319,7 @@ class QwenImage21Transformer2DModel(nn.Module):
         mod = (
             _split_rows(scale1), _split_rows(mx.tanh(gate1)),
             _split_rows(scale2), _split_rows(mx.tanh(gate2)),
-            mx.zeros((1, 1, scale1.shape[-1])),
+            mx.zeros((1, 1, scale1.shape[-1]), dtype=dtype),
         )
 
         attn_fn = block_causal_attention(segments)
